feat: enhance search feature
