+++ packages/core/src/web/app/views/beambox/Right-Panels/Options-Blocks/TextOptions/components/GoogleFontsPanel.module.scss
@@ -70,25 +70,79 @@
   .searchSection {
     margin-bottom: 32px;
 
-    .searchInput {
-      width: 100%;
+    .searchInputContainer {
+      position: relative;
 
-      .ant-input {
-        border-radius: 24px;
-        padding: 8px 16px 8px 40px;
-        border: 1px solid #dadce0;
-        background: #f8f9fa;
+      .searchInput {
+        width: 100%;
 
-        &:focus {
-          border-color: #1a73e8;
-          box-shadow: 0 0 0 1px #1a73e8;
-          background: #fff;
+        .ant-input {
+          border-radius: 24px;
+          padding: 8px 16px 8px 40px;
+          border: 1px solid #dadce0;
+          background: #f8f9fa;
+
+          &:focus {
+            border-color: #1a73e8;
+            box-shadow: 0 0 0 1px #1a73e8;
+            background: #fff;
+          }
+        }
+
+        .ant-input-prefix {
+          margin-left: 8px;
+          color: #5f6368;
         }
       }
 
-      .ant-input-prefix {
-        margin-left: 8px;
-        color: #5f6368;
+      .searchSuggestions {
+        position: absolute;
+        top: 100%;
+        left: 0;
+        right: 0;
+        background: #fff;
+        border: 1px solid #dadce0;
+        border-radius: 8px;
+        box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+        z-index: 1000;
+        max-height: 240px;
+        overflow-y: auto;
+        margin-top: 4px;
+
+        .suggestionItem {
+          display: flex;
+          align-items: center;
+          gap: 8px;
+          padding: 12px 16px;
+          cursor: pointer;
+          transition: background-color 0.2s ease;
+          border-bottom: 1px solid #f1f3f4;
+
+          &:last-child {
+            border-bottom: none;
+          }
+
+          &:hover,
+          &.highlighted {
+            background: #f8f9fa;
+          }
+
+          &.highlighted {
+            background: #e8f0fe;
+          }
+
+          .suggestionIcon {
+            color: #5f6368;
+            font-size: 14px;
+            flex-shrink: 0;
+          }
+
+          span {
+            color: #3c4043;
+            font-size: 14px;
+            font-weight: 400;
+          }
+        }
       }
     }
   }
